requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
spacy==3.7.5
python-dotenv==1.0.1